Restructure all api ts files;
