+++ src/api/common.ts
@@ -1,0 +1,8 @@
+export interface ApiResponse<T> {
+  status: string;
+  detail: string;
+  data: T | null | undefined;
+}
+
+export const apiPrefix =
+  import.meta.env.MODE == "development" ? "http://localhost:8080" : "";
