+++ src/components/UserProfile.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
-import { decodeJwt } from "../common";
+import { UserApi } from "../api/user";
 
 const jwt = localStorage.getItem("jwt")
-  ? decodeJwt(localStorage.getItem("jwt") as string)
+  ? UserApi.decodeJwt(localStorage.getItem("jwt") as string)
   : null;
 const email = jwt ? jwt.email : null;
 </script>
